All issues with merge fixed
Fixed all missing objects in main and edited some script to clear bugs. Bugs include: Event text for announcement not being cleared the next day. Storage amount adding on top of itself instead of replacing amount. Summary totals not being called in storefront.
Removed upkeep and downgrade feature as would require redesign of object to reflect the info and the game is not balanced for that.